requirements.txt
beautifulsoup4
requests
feedparser
pandas
openai
python-dotenv
alpha_vantage
yfinance
timedelta